refactor: created actions file for concerts page
